Added file structure to GUI

+++ Prototype/src/System/GUI/UIElement/UIBorder.cpp
@@ -11,6 +11,6 @@ UIBorder::UIBorder(int _posX, int _posY, int _border_width, int _border_height, 
 void UIBorder::render()
 {
     SDL_SetRenderDrawColor(GraphicsEngine::getActiveRenderer(), color.r, color.g, color.b, color.a);
-    SDL_Rect rect = { positionX, positionY, width, height };
+    SDL_Rect rect = { positionX, positionY, (int)width, (int)height };
     SDL_RenderFillRect(GraphicsEngine::getActiveRenderer(), &rect);
 }
